+++ src/infrastructure/handler/grpc/auth/mapper/mapper.go
@@ -9,7 +9,7 @@ type Mapper struct {
 }
 
 func (m Mapper) ToDomainRegisterRequest(registerReq *pb.RegisterRequest) *user.User {
-	return user.NewUser(0, registerReq.Name, registerReq.UserName, registerReq.Email, registerReq.Password, "", m.toDomainRole(registerReq.Role), true)
+	return user.NewUser(0, registerReq.Name, registerReq.UserName, registerReq.Email, registerReq.Password, "", m.toDomainRole(registerReq.Role), true, []user.Notification{})
 }
 
 func (m Mapper) ToDTORegisterResponse(domainUser *user.User) *pb.RegisterResponse {
@@ -17,7 +17,7 @@ func (m Mapper) ToDTORegisterResponse(domainUser *user.User) *pb.RegisterRespons
 }
 
 func (m Mapper) ToDomainLoginRequest(loginReq *pb.LoginRequest) *user.User {
-	return user.NewUser(0, "", loginReq.UserName, "", loginReq.Password, "", "", true)
+	return user.NewUser(0, "", loginReq.UserName, "", loginReq.Password, "", "", true, []user.Notification{})
 }
 
 func (m Mapper) ToDTOLoginResponse(domainUser *user.User) *pb.LoginResponse {
@@ -30,7 +30,7 @@ func (m Mapper) ToDTOLoginResponse(domainUser *user.User) *pb.LoginResponse {
 }
 
 func (m Mapper) ToDomainValidateRequest(validateReq *pb.ValidateRequest) *user.User {
-	return user.NewUser(0, "", "", "", "", validateReq.Token, "", true)
+	return user.NewUser(0, "", "", "", "", validateReq.Token, "", true, []user.Notification{})
 }
 
 func (m Mapper) ToDTOValidateResponse(domainUser *user.User) *pb.ValidateResponse {
@@ -59,7 +59,29 @@ func (m Mapper) ToDTOListResponse(domainUsers *[]user.User) *pb.ListResponse {
 
 func (m Mapper) ToDomainUpdateRequest(updateReq *pb.UpdateRequest) *user.User {
 	dto := updateReq.User
-	return user.NewUser(dto.UserId, dto.Name, dto.UserName, dto.Email, dto.Password, "", m.toDomainRole(dto.Role), dto.Status)
+	return user.NewUser(dto.UserId, dto.Name, dto.UserName, dto.Email, dto.Password, "", m.toDomainRole(dto.Role), dto.Status, []user.Notification{})
+}
+
+func (m Mapper) ToDTOCurrentResponse(domainUser *user.User) *pb.CurrentResponse {
+	var dto pb.CurrentResponse
+	for _, domainNotif := range domainUser.Notifications() {
+		tmp := pb.CurrentResponse_UserNotification{
+			Id:     domainNotif.Id(),
+			Title:  domainNotif.Title(),
+			Detail: domainNotif.Detail(),
+			Date:   domainNotif.Date().String(),
+			Read:   domainNotif.Read(),
+		}
+		dto.Notifications = append(dto.Notifications, &tmp)
+	}
+
+	dto.UserId = domainUser.Id()
+	dto.Name = domainUser.Name()
+	dto.UserName = domainUser.UserName()
+	dto.Email = domainUser.Email()
+	dto.Role = m.toDTORole(domainUser.Role())
+
+	return &dto
 }
 
 func (m Mapper) toDomainRole(dtoRole string) user.Role {