+++ src/infrastructure/platform/postgresql/client.go
@@ -1,7 +1,6 @@
 package postgresql
 
 import (
-	"github.com/lumialvarez/go-grpc-auth-service/src/infrastructure/repository/postgresql/user/dao"
 	"log"
 
 	"gorm.io/driver/postgres"
@@ -20,8 +19,6 @@ func Init(url string) Client {
 	}
 
 	log.Print("Connected to " + url)
-
-	db.AutoMigrate(dao.User{})
 
 	return Client{db}
 }