+++ src/infrastructure/repository/postgresql/user/mapper/mapper.go
@@ -9,7 +9,13 @@ type Mapper struct {
 }
 
 func (m Mapper) ToDomain(daoUser *dao.User) *user.User {
-	return user.NewUser(daoUser.Id, daoUser.Name, daoUser.UserName, daoUser.Email, daoUser.Password, "", user.Role(daoUser.Rol), daoUser.Status)
+	var domainNotifications []user.Notification
+	for _, daoNotification := range daoUser.UserNotification {
+		domainNotification := user.NewNotification(daoNotification.Id, daoNotification.Title, daoNotification.Detail, daoNotification.Date, daoNotification.Read)
+		domainNotifications = append(domainNotifications, *domainNotification)
+	}
+
+	return user.NewUser(daoUser.Id, daoUser.Name, daoUser.UserName, daoUser.Email, daoUser.Password, "", user.Role(daoUser.Rol), daoUser.Status, domainNotifications)
 }
 
 func (m Mapper) ToDAO(domainUser *user.User) *dao.User {